% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAT}
\alias{CAT}
\title{Pseudo activity count (per minute) data for cats}
\format{A data frame with 4320 rows and 5 variables:
\describe{
  \item{id}{cat ID: 1,2,3}
  \item{hour}{hour of the day: 1,2,...,24}
  \item{minute}{minute of the hour: 1,2,...,60}
  \item{night}{night time indicator}
  \item{activity}{activity count data}
}}
\usage{
CAT
}
\description{
A simulated dataset containing the per minute activity count for cats
}
\keyword{datasets}

