% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolution}
\alias{convolution}
\title{Convolution of two real vectors of the same length.}
\usage{
convolution(vec1, vec2)
}
\arguments{
\item{vec1}{the first vector}

\item{vec2}{the second vector}
}
\value{
a vector of full convolution
}
\description{
Convolution of two real vectors of the same length.
}
