% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dzip}
\alias{dzip}
\title{pmf for zero-inflated poisson}
\usage{
dzip(p, theta, y, loga)
}
\arguments{
\item{p}{proportion of structural zero's}

\item{theta}{the poisson mean}

\item{y}{the observed value}

\item{loga}{Logical. Whether to return the log probability or not.}
}
\value{
the probability mass of the zero-inflated poisson distribution
}
\description{
pmf for zero-inflated poisson
}
