% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_zipnegloglik_nocov_cont}
\alias{grad_zipnegloglik_nocov_cont}
\title{gradient for negative log likelihood function from zero-inflated Poisson hidden Markov model
without covariates, where zero-inflation only happens in state 1}
\usage{
grad_zipnegloglik_nocov_cont(parm, M, y, ntimes, timeindex, udiff)
}
\arguments{
\item{parm}{working parameters}

\item{M}{number of hidden states}

\item{y}{observed series}

\item{ntimes}{length of the observed series}

\item{timeindex}{vector of observed time points}

\item{udiff}{unique time intervals}
}
\value{
gradient for negative log likelihood
}
\description{
gradient for negative log likelihood function from zero-inflated Poisson hidden Markov model
without covariates, where zero-inflation only happens in state 1
}
