% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package-ziphsmm}
\alias{package-ziphsmm}
\alias{ziphsmm-package}
\alias{package-ziphsmm-package}
\title{zero-inflated poisson hidden (semi-)Markov models}
\description{
\tabular{ll}{
Package: \tab ziphsmm \cr
Type: \tab Package\cr
Version: \tab 2.0.6\cr
Date: \tab 2018-05-08\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
Zekun Xu \email{zekunxu@gmail.com}

Ye Liu
Maintainer: Zekun Xu \email{zekunxu@gmail.com}
}
\keyword{Markov}
\keyword{hidden}
\keyword{models}
\keyword{poisson,}
\keyword{zero-inflated}
