% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_cov_cont3}
\alias{retrieve_cov_cont3}
\title{retrieve the natural parameters from the working parameters in zero-inflated Poisson
hidden Markov model with covariates in state-dependent parameters and transition rates}
\usage{
retrieve_cov_cont3(parm, M, ncolx)
}
\arguments{
\item{parm}{working parameters}

\item{M}{number of hidden states}

\item{ncolx}{number of covariates including the intercept}
}
\value{
a list of natural parameters
}
\description{
retrieve the natural parameters from the working parameters in zero-inflated Poisson
hidden Markov model with covariates in state-dependent parameters and transition rates
}
