% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_nocov_cont}
\alias{retrieve_nocov_cont}
\title{retrieve the natural parameters from working parameters for a continuous-time
zero-inflated Poisson hidden Markov model where zero-inflation only happens in state 1}
\usage{
retrieve_nocov_cont(parm, M)
}
\arguments{
\item{parm}{working parameters}

\item{M}{number of hidden states}
}
\value{
a list of natural parameters
}
\description{
retrieve the natural parameters from working parameters for a continuous-time
zero-inflated Poisson hidden Markov model where zero-inflation only happens in state 1
}
