% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rzip}
\alias{rzip}
\title{generate zero-inflated poisson random variables}
\usage{
rzip(n, p, theta)
}
\arguments{
\item{n}{length of the random series}

\item{p}{proportion of structural zero's}

\item{theta}{the poisson mean}
}
\value{
a series of zero-inflated poisson random variables
}
\description{
generate zero-inflated poisson random variables
}
