% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zipnegloglik_cov_cont}
\alias{zipnegloglik_cov_cont}
\title{negative log likelihood function for zero-inflated Poisson hidden Markov model with covariates,
where zero-inflation only happens in state 1}
\usage{
zipnegloglik_cov_cont(parm, y, covariates, M, ntimes, timeindex, udiff)
}
\arguments{
\item{parm}{working parameters}

\item{y}{observed series}

\item{covariates}{design matrix of covariates including the intercept}

\item{M}{number of hidden states}

\item{ntimes}{length of the observed series}

\item{timeindex}{vector of observed time points}

\item{udiff}{unique time intervals}
}
\value{
negative log likelihood
}
\description{
negative log likelihood function for zero-inflated Poisson hidden Markov model with covariates,
where zero-inflation only happens in state 1
}
