% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zipnegloglik_cov_cont3}
\alias{zipnegloglik_cov_cont3}
\title{negative log likelihood function for zero-inflated Poisson hidden Markov model
with covariates in state-dependent parameters and transition rates}
\usage{
zipnegloglik_cov_cont3(parm, y, covariates, M, ntimes, timeindex)
}
\arguments{
\item{parm}{working parameters}

\item{y}{observed series}

\item{covariates}{design matrix of covariates including the intercept}

\item{M}{number of hidden states}

\item{ntimes}{length of the observed series}

\item{timeindex}{vector of observed time points}
}
\value{
negative log likelihood
}
\description{
negative log likelihood function for zero-inflated Poisson hidden Markov model
with covariates in state-dependent parameters and transition rates
}
